--- Cpp_Files/output2.txt
@@ -1,30 +1,0 @@
-========================================
-           SeedingBot Log
-========================================
-  Seeds Planted: 1
-  Seed Log:
-    Session 1 | Spinach | Time to Grow  : 5 days | Water/day     : 4.00 mL | Humidity      : 60 - 75 % | Temperature   : 15 - 25 C | UV Intensity  : 10 - 20
-
-========================================
-           SprayerBot Log
-========================================
-  Sessions Done: 0
-  Spray Log:
-    (none)
-
-========================================
-          HarvestingBot Log
-========================================
-  Harvests Done  : 1
-  Total Harvested: 0.98 kg
-    Plant        : 0.98 kg
-    Dead         : 0.00 kg
-  Harvest Log:
-    [Plant Harvests]
-      Spinach | 0.98 kg
-    [Dead Harvests]
-      (none)
-
-========================================
-            End of Report
-========================================
